+++ silo-core/contracts/lib/SiloSolvencyLib.sol
@@ -139,7 +139,7 @@ library SiloSolvencyLib {
 
         uint256 totalDebtAssets = _accrueInMemory == ISilo.AccrueInterestInMemory.Yes
             ? SiloStdLib.getTotalDebtAssetsWithInterest(_debtConfig.silo, _debtConfig.interestRateModel)
-            : ISilo(_debtConfig.silo).total(uint8(ISilo.AssetType.Debt));
+            : ISilo(_debtConfig.silo).total(uint256(ISilo.AssetType.Debt));
 
         // BORROW value -> to assets -> UP
         ltvData.borrowerDebtAssets = SiloMathLib.convertToAssets(
