+++ silo-core/test/echidna/internal_testing/SiloInternal.sol
@@ -14,12 +14,12 @@ contract SiloInternal is Silo, CryticIERC4626Internal {
     function recognizeProfit(uint256 profit) public {
         address _asset = sharedStorage.siloConfig.getAssetForSilo(address(this));
         TestERC20Token(address(_asset)).mint(address(this), profit);
-        total[uint8(ISilo.AssetType.Collateral)].assets += profit;
+        total[uint256(ISilo.AssetType.Collateral)].assets += profit;
     }
 
     function recognizeLoss(uint256 loss) public {
         address _asset = sharedStorage.siloConfig.getAssetForSilo(address(this));
         TestERC20Token(address(_asset)).burn(address(this), loss);
-        total[uint8(ISilo.AssetType.Collateral)].assets -= loss;
+        total[uint256(ISilo.AssetType.Collateral)].assets -= loss;
     }
 }
