+++ silo-core/test/foundry/Silo/deposit/Deposit.i.sol
@@ -76,7 +76,7 @@ contract DepositTest is SiloLittleHelper, Test {
 
         assertEq(token0.balanceOf(address(silo0)), assets * 2);
         assertEq(silo0.getCollateralAssets(), assets);
-        assertEq(silo0.total(uint8(ISilo.AssetType.Protected)), assets);
+        assertEq(silo0.total(uint256(ISilo.AssetType.Protected)), assets);
         assertEq(silo0.getDebtAssets(), 0);
 
         assertEq(IShareToken(collateral.collateralShareToken).balanceOf(depositor), assets, "collateral shares");
@@ -84,7 +84,7 @@ contract DepositTest is SiloLittleHelper, Test {
 
         assertEq(token1.balanceOf(address(silo1)), assets * 2);
         assertEq(silo1.getCollateralAssets(), assets);
-        assertEq(silo1.total(uint8(ISilo.AssetType.Protected)), assets);
+        assertEq(silo1.total(uint256(ISilo.AssetType.Protected)), assets);
         assertEq(silo1.getDebtAssets(), 0);
 
         assertEq(IShareToken(debt.collateralShareToken).balanceOf(depositor), assets, "collateral shares (on other silo)");
